refactor(python): move channel options to config.py with get_channel_options() extension point
